Refactor server.js with async/await

+++ server/server.js
@@ -65,24 +65,23 @@ app.get('/todos/:id', authenticate, (req, res) => {
 });
 
 // DELETE /todos
-app.delete('/todos/:id', authenticate, (req, res) => {
-  let id = req.params.id;
+app.delete('/todos/:id', authenticate, async (req, res) => {
+  const id = req.params.id;
 
   if (!ObjectID.isValid(id)) {
     return res.status(404).send();
   }
 
-  Todo.findOneAndRemove({
-    _id: id,
-    _creator: req.user._id
-  }).then((todo) => {
+  try {
+    const todo = await Todo.findOneAndRemove({ _id: id, _creator: req.user._id });
+
     if (!todo) {
       return res.status(404).send();
-    }
+    };
     res.send({todo});
-  }, (e) => {
-    res.status(400).send();
-  });
+  } catch (e) {
+    res.status(400).send(e);  
+  }
 
 });
 
@@ -118,17 +117,16 @@ app.patch('/todos/:id', authenticate, (req, res) => {
 });
 
 // POST /users
-app.post('/users', (req, res) => {
+app.post('/users', async (req, res) => {
 
-  let user = new User(_.pick(req.body, ['email', 'password']));
-
-  user.save().then(() => {
-    return user.generateAuthToken();
-  }).then((token) => {
+  try {
+    const user = new User(_.pick(req.body, ['email', 'password']));
+    await user.save();
+    const token = await user.generateAuthToken();
     res.header('x-auth', token).send(user);
-  }).catch((e) => {
+  } catch (e) {
     res.status(400).send(e);
-  });
+  };
 
 });
